Move currencyRound function to own file
